moving params list to list params to be consistent with verb noun and to add a number of other list types for operators
